moves controllers to containers following the todomvc example
