+++ src/entry.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import App from './controllers/App.jsx';
+import App from './containers/App.jsx';
 import configureStore from './store/configureStore';
 
 const store = configureStore();
